% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{somar}
\alias{somar}
\title{Função Somar
A função realiza a soma de dois números.}
\usage{
somar(a, b)
}
\arguments{
\item{a}{recebe a primeira parcela}

\item{b}{recebe a segunda parcela}
}
\description{
Função Somar
A função realiza a soma de dois números.
}
\examples{
somar(3,4)
}
