% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtracao.R
\name{subtrair}
\alias{subtrair}
\title{Função Subtrair
A função realiza a subtraição de dois números.}
\usage{
subtrair(a, b)
}
\arguments{
\item{a}{recebe a primeira parcela}

\item{b}{recebe a segunda parcela}
}
\description{
Função Subtrair
A função realiza a subtraição de dois números.
}
\examples{
subtrair(3,4)
}
